Add FullScreenImageView component and integrate with ImageDisplayer; update styles for better presentation
