+++ client/src/components/Chat/ConversationScreen/ChatMessages.css
@@ -104,7 +104,7 @@
     width: 100%;
     height: 300px;
     object-fit: cover;
-    border-radius: 5px;
+    border-radius: 10px;
 }
 
 .chat-messages .row .message .message-media.media-doc {
