+++ client/src/components/FullScreenImageView.css
@@ -1,0 +1,14 @@
+#full-screen-image-viewer {
+    width: 80%;
+    max-height: 80%;
+    background-color: #fff;
+    height: 80%;
+    border-radius: 5px;
+    overflow: hidden;
+}
+
+#full-screen-image-viewer img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
