+++ client/src/components/Support/TicketDetailView.css
@@ -130,18 +130,3 @@
 .support-ticket-comment:hover {
     background-color: #fafafa;
 }
-
-#full-screen-image-viewer {
-    width: 80%;
-    max-height: 80%;
-    background-color: #fff;
-    height: 80%;
-    border-radius: 5px;
-    overflow: hidden;
-}
-
-#full-screen-image-viewer img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
